+++ CHANGELOG.md
@@ -6,7 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Fixed
+- [CU-2azm63y] module Schedule Delivery by [kienluu612@gmail.com].
+### Changed
+- Re-design android wheel picker by [kienluu612@gmail.com].
 ## [r13.1.9] - 2022-05-30
 ### Changed
 - [CU-2d9zj99] reset app when user changing sale point by [kienluu612@gmail.com].